Added cli options for throttling files and center frequency to support a fixed freq range type

+++ scanner/src/process/context.rs
@@ -4,8 +4,8 @@ use crate::io::Output;
 use crate::process::SignalMetadata;
 
 pub struct ProcessContext {
-    pub(crate) center_freq: f32,
-    pub(crate) sample_rate: f32,
+    pub(crate) center_freq: u32,
+    pub(crate) sample_rate: u32,
     pub(crate) process_type: ProcessType,
     pub(crate) out_tx: Sender<Output>,
     pub metadata: SignalMetadata
@@ -13,8 +13,8 @@ pub struct ProcessContext {
 
 impl ProcessContext {
     pub fn new(
-        center_freq: f32,
-        sample_rate: f32,
+        center_freq: u32,
+        sample_rate: u32,
         process_type: ProcessType,
         out_tx: Sender<Output>,
     ) -> Self {
